refactor: organize components files based on domains
