+++ src/components/auth/index.tsx
@@ -1,15 +1,25 @@
-import EmailVerificationForm from "@/components/auth/forms/email-verification";
-import PasswordSetupForm from "@/components/auth/forms/password-setup";
-import PhoneLoginForm from "@/components/auth/forms/phone-login";
-import PhoneOtpLoginForm from "@/components/auth/forms/phone-otp-login";
-import PhoneVerificationForm from "@/components/auth/forms/phone-verification";
-import SignUpModal from "@/components/auth/sign-up-modal";
+/**
+ * Authentication Components
+ *
+ * Components for user authentication including sign-up, sign-in,
+ * and social account management.
+ */
 
-export default SignUpModal;
-export {
-  EmailVerificationForm,
-  PasswordSetupForm,
-  PhoneLoginForm,
-  PhoneOtpLoginForm,
-  PhoneVerificationForm,
-};
+// Forms
+export { default as EmailLoginForm } from "./forms/email-login";
+export { default as EmailOtpLoginForm } from "./forms/email-otp-login";
+export { default as EmailVerificationForm } from "./forms/email-verification";
+export { default as PasswordSetupForm } from "./forms/password-setup";
+export { default as PhoneLoginForm } from "./forms/phone-login";
+export { default as PhoneOtpLoginForm } from "./forms/phone-otp-login";
+export { default as PhoneVerificationForm } from "./forms/phone-verification";
+
+// Other components
+export { LinkButton } from "./link-button";
+export { default as SignUpModal } from "./sign-up-modal";
+export { default as SocialAccountCard } from "./social-account-card";
+export { UnlinkButton } from "./unlink-button";
+export { default as UserSidebar } from "./user-sidebar";
+
+// Default export
+export { default } from "./sign-up-modal";
