+++ src/components/flights/booking/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Flight Booking Components
+ *
+ * Components for the flight booking process including passenger forms,
+ * contact information, flight summary, and ancillary services.
+ */
+
+export { default as AncillarySelection } from "./ancillary-selection";
+export { default as ContactInfoCard } from "./contact-info-card";
+export { default as FlightSummaryCard } from "./flight-summary-card";
+export { default as PassengerFormCard } from "./passenger-form-card";
