+++ src/components/flights/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Flight Components
+ *
+ * All flight-related components organized by functionality:
+ * - search: Flight search forms and selectors
+ * - results: Flight search results display
+ * - booking: Flight booking process
+ * - orders: Order management and display
+ */
+
+// Re-export all search components
+export * from "./search";
+
+// Re-export all results components
+export * from "./results";
+
+// Re-export all booking components
+export * from "./booking";
+
+// Re-export all order components
+export * from "./orders";
