+++ src/components/flights/results/flight-card-skeleton.tsx
