+++ src/components/flights/search/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Flight Search Components
+ *
+ * Components related to flight search functionality including
+ * search forms, city/date selectors, and search history.
+ */
+
+export { default as CitySelector } from "./city-selector";
+export { default as DateSelector } from "./date-selector";
+export { default as QuickDateSelector } from "./quick-date-selector";
+export { default as SearchForm } from "./search-form";
+export { default as SearchHistory } from "./search-history";
