+++ src/components/security/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Security Settings Components
+ *
+ * Components for managing user security settings including
+ * password changes and contact information updates.
+ */
+
+export { default as ChangePasswordForm } from "./change-password-form";
+export { default as UpdateEmailForm } from "./update-email-form";
+export { default as UpdatePhoneForm } from "./update-phone-form";
