+++ src/hooks/use-passenger-forms.test.ts
@@ -1,7 +1,7 @@
 import { act, renderHook } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
 
-import type { PassengerFormData } from "@/components/flights/passenger-form-card";
+import type { PassengerFormData } from "@/components/flights/booking/passenger-form-card";
 
 import {
   isFormEmpty,
