+++ src/stories/auth/forms/password-setup-form.stories.tsx
