+++ src/stories/flights/booking/contact-info-card.stories.tsx
@@ -5,7 +5,7 @@ import {
   type ContactInfo,
   ContactInfoCard,
   type ContactInfoValidationErrors,
-} from "@/components/flights/contact-info-card";
+} from "@/components/flights/booking/contact-info-card";
 
 const meta: Meta<typeof ContactInfoCard> = {
   title: "Flights/ContactInfoCard",
